test and fix If-Match handling
Seeking in Chrome 55 wasn't working. It apparently sends If-Match requests
with the correct etag, which Moonfire NVR was incorrectly responding to with
"Precondition failed" responses. Fix and test that.

I hadn't noticed the problem in earlier versions of Chrome. I think they were
using If-Range instead, which is already tested and working.